// src/app/models/admin/dangkyhoatdongkhoangsan/dangkydaugia/dangkydaugia.model.ts
export class OutputDKDauGiaKhoangSanModel {
  iddangkydaugia: string;
  idthongbao: string;
  sothongbao: string;
  ngaythongbao: Date;
  coquanthongbao: string;
  tenkhuvucdaugia: string;
  diadiem: string;
  tenloaikhoangsan: string;
  dientichkhuvuc: number;
  donvidientich: string;
  idhoso: string;
  siteid: string;
}

export class InputDKDauGiaKhoangSanModel {
  iddangkydaugia: string;
  idthongbao: string;
  sothongbao: string;
  ngaythongbao: Date;
  coquanthongbao: string;
  tenkhuvucdaugia: string;
  diadiem: string;
  tenloaikhoangsan: string;
  dientichkhuvuc: number;
  donvidientich: string;
  idhoso: string;
  siteid: string;
}